requirements.txt
aiohttp
beautifulsoup4
fake-useragent
google-search-results
openai
pydantic 
pypdf
python-dotenv
trafilatura